+++ front/src/type/QA-type.ts
@@ -1,0 +1,4 @@
+export interface QA {
+  query: string;
+  answer: string;
+}
